requirements-semantic.txt
# Semantic Processing Dependencies
sentence-transformers==2.2.2
transformers==4.21.0
torch==1.12.0
scikit-learn==1.1.0
numpy==1.23.0

# Existing Dependencies
fastapi==0.68.0
uvicorn==0.15.0
streamlit==1.12.0
pandas==1.4.3
PyPDF2==2.10.0
python-docx==0.8.11
sqlalchemy==1.4.39
psycopg2-binary==2.9.3
httpx==0.24.0
python-multipart==0.0.5